requirements.txt
flask
python-secrets
azure-cosmos
